Disable signal handling of JVM GC signals
The JVM uses signals as an efficient way to do GC safe-pointing. When
clang is loaded as a native library on the JVM, the LLVM signal handler
ends up getting triggered by the signals generated by the GC mechanism.
This problem is addressed by libjsig, but that approach requires
preloading the library on JVM startup, which cannot be ensured when
using Gradle daemons. Furthermore, libjsig does not work on newer
versions of MacOS.